bot/requirements.txt
python-telegram-bot[all]==21.0
qrcode[pil]==7.4.2
pillow==10.1.0
requests==2.31.0
